Don't attempt to delete ruleset from "other" trees
When a ruleset was being applied to an interface, the code previously
attempted to find out if that ruleset name was being applied to that
same interface and in the same direction, but under a different "tree"
name (e.g. "name", "ipv6-name", "modify", etc.).  If it found a match, it
would delete the other rule.  But the matching logic
was insufficient, so it killed off some random other rule instead.

There is really no need to perform this check anyway, so I have
removed it.