network-groups: T4869: Fixed operations with /32 and /128 netmasks
When the configuration script performs operations with network-group items
received from CLI, it gets them in the format fixed by CLI restrictions - always
with netmasks. But inside ipset networks with netmasks /32 for IPv4 and
/128 for IPv6 for some reason represented as items without netmask. This breaks
comparison logic in the configuration script that relies on a direct match
between items.

This commit adds extra normalization for data received from ipset - an
appropriate netmask is added to networks with /32 and /128 netmasks while
preparing a hash with items. This allows using hash keys for matching as it is
intended to be.